requirements.txt
streamlit 
beautifulsoup4
lxml 
html5lib
python-dotenv
google-generativeai
fake-useragent
requests
pandas
plotly
selenium
praw
webdriver-manager